+++ Sources/AlgoliaSearchClient/Client/Search/SearchClient.swift
@@ -77,6 +77,18 @@ extension SearchClient: TransportContainer {}
 
 extension SearchClient {
 
+  func execute<Output: Codable & AppTask>(_ command: AlgoliaCommand, completion: @escaping ResultAppTaskCallback<Output>) -> Operation & TransportTask {
+    transport.execute(command, transform: WaitableWrapper.wrap(with: self), completion: completion)
+  }
+
+  func execute<Output: Codable & AppTask>(_ command: AlgoliaCommand) throws -> WaitableWrapper<Output> {
+    try transport.execute(command, transform: WaitableWrapper.wrap(with: self))
+  }
+
+}
+
+extension SearchClient {
+
   @discardableResult func launch<O: Operation>(_ operation: O) -> O {
     return operationLauncher.launch(operation)
   }