+++ Sources/AlgoliaSearchClient/Helpers/ResultCallback.swift
@@ -9,4 +9,5 @@ import Foundation
 
 public typealias ResultCallback<T> = (Result<T, Error>) -> Void
 public typealias ResultTaskCallback<T: Task & Codable> = (Result<WaitableWrapper<T>, Error>) -> Void
+public typealias ResultAppTaskCallback<T: AppTask & Codable> = (Result<WaitableWrapper<T>, Error>) -> Void
 public typealias ResultBatchesCallback = (Result<WaitableWrapper<BatchesResponse>, Error>) -> Void