+++ Sources/AlgoliaSearchClient/Models/Settings/SettingsParametersCodingKeys.swift
@@ -34,10 +34,12 @@ public enum SettingsParametersCodingKeys: String, CodingKey {
   case separatorsToIndex
   case ignorePlurals
   case removeStopWords
+  case attributesToTransliterate
   case camelCaseAttributes
   case decompoundedAttributes
   case keepDiacriticsOnCharacters
   case queryLanguages
+  case decompoundQuery
   case enableRules
   case enablePersonalization
   case queryType