+++ Tests/AlgoliaSearchClientTests/Doc/APIParameters/APIParameters+Languages.swift
@@ -246,5 +246,47 @@ extension APIParameters {
         }
       }
     }
+    func attributesToTransliterate() {
+      /*
+       attributesToTransliterate = ["name", "description"]
+       */
+      func set_attributes_to_transliterate() {
+        let settings = Settings()
+          .set(\.attributesToTransliterate, to: ["name", "description"])
+
+        index.setSettings(settings) { result in
+          if case .success(let response) = result {
+            print("Response: \(response)")
+          }
+        }
+      }
+    }
+    func decompoundQuery() {
+      /*
+       decompoundQuery = true
+       */
+      func enable_decompound_query_by_default() {
+        let settings = Settings()
+          .set(\.decompoundQuery, to: true)
+        
+        index.setSettings(settings) { result in
+          if case .success(let response) = result {
+            print("Response: \(response)")
+          }
+        }
+      }
+      
+      func enable_decompound_query_search_time() {
+        let query = Query("query")
+          .set(\.decompoundQuery, to: true)
+        
+        index.search(query: query) { result in
+          if case .success(let response) = result {
+            print("Response: \(response)")
+          }
+        }
+      }
+    }
+
   }
 }